Reimplement and finish Day 7 using ubergraph
I was more interested in experimenting with ubergraph in this solution than
making it efficient, so this could easily be optimized and be made more elegant.
The inefficiencies don't lead to any noticeable slowdown on my machine for input
of this size, but I might go back at some point to improve performance as more
of an academic exercise.