Add Day 15 solution
I've finally read up on transducers and reducers to my great joy. I originally
rolled a stateful transducer from scratch for this problem for practice but then
replaced it with the built-in map and take solution now seen here.

There's a couple easy refactors to the transducer-builder functions used here
and `count-=-xforms` that would simplify how the latter is invoked, but those
can wait.